Refactors index only commands
Extract out all commands that only need 1 index parameter to ensure that their functionality is unified
