+++ src/main/java/duke/Utils.java
@@ -8,6 +8,7 @@ import java.time.format.DateTimeParseException;
 import java.time.temporal.TemporalAccessor;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Locale;
 import java.util.Optional;
 import java.util.function.BiFunction;
 import java.util.function.Function;
@@ -24,9 +25,9 @@ public final class Utils {
     return (str) -> {
         for (String pattern : patterns) {
             try {
-                return Optional.of(parser.apply(str, DateTimeFormatter.ofPattern(pattern)));
+                return Optional.of(parser.apply(str, DateTimeFormatter.ofPattern(pattern, Locale.ENGLISH)));
             } catch (DateTimeParseException r) {
-              System.out.format("Couldn't match %s with %s\n", str, pattern);
+              // System.out.format("Couldn't match %s with %s\n", str, pattern);
             }
         }
         return Optional.empty();
